Adress PR review
