+++ sentry-android-core/src/test/java/io/sentry/android/core/AndroidOptionsInitializerTest.kt
@@ -55,13 +55,13 @@ class AndroidOptionsInitializerTest {
             assets: AssetManager? = null
         ) {
             mockContext = if (metadata != null) {
-                ContextUtilsHelper.mockMetaData(
-                    mockContext = ContextUtilsHelper.createMockContext(hasAppContext),
+                ContextUtilsTestHelper.mockMetaData(
+                    mockContext = ContextUtilsTestHelper.createMockContext(hasAppContext),
                     metaData = metadata,
                     assets = assets
                 )
             } else {
-                ContextUtilsHelper.createMockContext(hasAppContext)
+                ContextUtilsTestHelper.createMockContext(hasAppContext)
             }
             whenever(mockContext.cacheDir).thenReturn(file)
             if (mockContext.applicationContext != null) {
@@ -101,8 +101,8 @@ class AndroidOptionsInitializerTest {
             isFragmentAvailable: Boolean = false,
             isTimberAvailable: Boolean = false
         ) {
-            mockContext = ContextUtilsHelper.mockMetaData(
-                mockContext = ContextUtilsHelper.createMockContext(hasAppContext = true),
+            mockContext = ContextUtilsTestHelper.mockMetaData(
+                mockContext = ContextUtilsTestHelper.createMockContext(hasAppContext = true),
                 metaData = Bundle().apply {
                     putString(ManifestMetadataReader.DSN, "https://key@sentry.io/123")
                 }
