+++ sentry-android-core/src/test/java/io/sentry/android/core/ManifestMetadataReaderTest.kt
@@ -30,7 +30,7 @@ class ManifestMetadataReaderTest {
         val buildInfoProvider = mock<BuildInfoProvider>()
 
         fun getContext(metaData: Bundle = Bundle()): Context {
-            return ContextUtilsHelper.mockMetaData(metaData = metaData)
+            return ContextUtilsTestHelper.mockMetaData(metaData = metaData)
         }
     }
