+++ sentry-android-core/src/test/java/io/sentry/android/core/SentryInitProviderTest.kt
@@ -26,7 +26,7 @@ class SentryInitProviderTest {
     fun `when missing applicationId, SentryInitProvider throws`() {
         val providerInfo = ProviderInfo()
 
-        val mockContext = ContextUtilsHelper.createMockContext()
+        val mockContext = ContextUtilsTestHelper.createMockContext()
         providerInfo.authority = SentryInitProvider::class.java.name
         assertFailsWith<IllegalStateException> { sentryInitProvider.attachInfo(mockContext, providerInfo) }
     }
@@ -39,7 +39,7 @@ class SentryInitProviderTest {
         providerInfo.authority = AUTHORITY
 
         val metaData = Bundle()
-        val mockContext = ContextUtilsHelper.mockMetaData(metaData = metaData)
+        val mockContext = ContextUtilsTestHelper.mockMetaData(metaData = metaData)
 
         metaData.putString(ManifestMetadataReader.DSN, "https://key@sentry.io/123")
 
@@ -56,7 +56,7 @@ class SentryInitProviderTest {
         providerInfo.authority = AUTHORITY
 
         val metaData = Bundle()
-        val mockContext = ContextUtilsHelper.mockMetaData(metaData = metaData)
+        val mockContext = ContextUtilsTestHelper.mockMetaData(metaData = metaData)
 
         metaData.putString(ManifestMetadataReader.DSN, "")
 
@@ -73,7 +73,7 @@ class SentryInitProviderTest {
         providerInfo.authority = AUTHORITY
 
         val metaData = Bundle()
-        val mockContext = ContextUtilsHelper.mockMetaData(metaData = metaData)
+        val mockContext = ContextUtilsTestHelper.mockMetaData(metaData = metaData)
 
         metaData.putString(ManifestMetadataReader.DSN, null)
 
@@ -88,7 +88,7 @@ class SentryInitProviderTest {
         providerInfo.authority = AUTHORITY
 
         val metaData = Bundle()
-        val mockContext = ContextUtilsHelper.mockMetaData(metaData = metaData)
+        val mockContext = ContextUtilsTestHelper.mockMetaData(metaData = metaData)
 
         metaData.putString(ManifestMetadataReader.DSN, "invalid dsn")
 
@@ -103,7 +103,7 @@ class SentryInitProviderTest {
         providerInfo.authority = AUTHORITY
 
         val metaData = Bundle()
-        val mockContext = ContextUtilsHelper.mockMetaData(metaData = metaData)
+        val mockContext = ContextUtilsTestHelper.mockMetaData(metaData = metaData)
 
         metaData.putBoolean(ManifestMetadataReader.AUTO_INIT, false)
 
@@ -129,7 +129,7 @@ class SentryInitProviderTest {
         val sentryOptions = SentryAndroidOptions()
 
         val metaData = Bundle()
-        val mockContext = ContextUtilsHelper.mockMetaData(metaData = metaData)
+        val mockContext = ContextUtilsTestHelper.mockMetaData(metaData = metaData)
         metaData.putBoolean(ManifestMetadataReader.NDK_ENABLE, false)
 
         AndroidOptionsInitializer.loadDefaultAndMetadataOptions(sentryOptions, mockContext)
